Refactor (moved Authorization files into the Authorization project except handlers
