+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Application/Commands/Authorization/CreateSignatureCommand.cs
@@ -13,6 +13,8 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MarketParticipant.Authorization;
+using Energinet.DataHub.MarketParticipant.Authorization.Restriction;
 using MediatR;
 
 namespace Energinet.DataHub.MarketParticipant.Application.Commands.Authorization;
